backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg[binary]==3.2.13
python-decouple==3.8
djangorestframework-simplejwt==5.3.1
dj-database-url==2.1.0
reportlab==4.4.10
openpyxl==3.1.5
apscheduler==3.10.4